+++ tests/unit/contentproviders/test_git.py
@@ -23,6 +23,18 @@ def test_submodule_clone(repo_with_submodule):
     """Test git clone containing a git submodule."""
     upstream, expected_sha1_upstream, expected_sha1_submod = repo_with_submodule
 
+    # check that checking out a branch where there are no submodule
+    # indeed doesn't get any submodule, even though they are in master
+    with TemporaryDirectory() as clone_dir2:
+        submod_dir = os.path.join(clone_dir2, "submod")  # set by fixture
+        spec = {"repo": upstream, "ref": "branch-without-submod"}
+        git_content = Git()
+        for _ in git_content.fetch(spec, clone_dir2):
+            pass
+
+        assert os.path.exists(os.path.join(clone_dir2, "test"))
+        assert not os.path.exists(os.path.join(submod_dir, "test"))
+
     with TemporaryDirectory() as clone_dir:
         submod_dir = os.path.join(clone_dir, "submod")  # set by fixture
         spec = {"repo": upstream}